Default options/params used in flowr and ngsflows

There are three helper functions which attempt to manage parameters used by flowr and ngsflows:

get_opts OR opts_flow\$get(): show all default options

set_opts OR opts_flow\$set(): set default options

load_opts OR opts_flow\$load(): load options specified in a tab separated text file

For more details regarding these functions refer to params package.

flowopts

get_opts(...)
set_opts(...)
load_opts(...)

Arguments

...

get: names of options to fetch

set: a set of options in a name=value format separated by commas

Format

opts_flow

An object of class list of length 5.

Details

By default flowr loads, ~/flowr/conf/flowr.conf and ~/flowr/conf/ngsflows.conf

Below is a list of default flowr options, retrieved via

opts_flow$get():

|name              |value                    |
|:-----------------|:------------------------|
|default_regex     |(.*)                     |
|flow_base_path    |~/flowr                  |
|flow_conf_path    |~/flowr/conf             |
|flow_parse_lsf    |.*(\<[0-9]*\>).*         |
|flow_parse_moab   |(.*)                     |
|flow_parse_sge    |(.*)                     |
|flow_parse_slurm  |(.*)                     |
|flow_parse_torque |(.?)\..*                 |
|flow_pipe_paths   |~/flowr/pipelines        |
|flow_pipe_urls    |~/flowr/pipelines        |
|flow_platform     |local                    |
|flow_run_path     |~/flowr/runs             |
|my_conf_path      |~/flowr/conf             |
|my_dir            |path/to/a/folder         |
|my_path           |~/flowr                  |
|my_tool_exe       |/usr/bin/ls              |
|time_format       |%a %b %e %H:%M:%S CDT %Y |
|verbose           |FALSE                    |

See also

Examples

## Set options: opts_flow$set()
opts = opts_flow$set(flow_run_path = "~/mypath")
## OR if you would like to supply a long list of options:
opts = opts_flow$set(.dots = list(flow_run_path = "~/mypath"))

## load options from a configuration file: opts_flow$load()
conffile = fetch_conf("flowr.conf")
opts_flow$load(conffile)

#> Reading file, using 'V1' as id_column to remove empty rows.

#> Warning: 
#> 
#> Seems like these paths do not exist, this may cause issues later:
#> |name           |value                   |
#> |:--------------|:-----------------------|
#> |flow_run_path  |~/flowr/runs            |
#> |flow_conf_path |{{flow_base_path}}/conf |
#> |flow_base_path |~/flowr                 |

## Fetch options: get_opts()
opts_flow$get("flow_run_path")

#>  flow_run_path 
#> "~/flowr/runs" 

opts_flow$get()

#> 
#> 
#> |name              |value                    |
#> |:-----------------|:------------------------|
#> |flow_base_path    |~/flowr                  |
#> |flow_conf_path    |{{flow_base_path}}/conf  |
#> |flow_parse_lsf    |.*(\<[0-9]*\>).*         |
#> |flow_parse_moab   |(.*)                     |
#> |flow_parse_sge    |(.*)                     |
#> |flow_parse_slurm  |.* ([0-9]+)              |
#> |flow_parse_torque |(.?)\..*                 |
#> |flow_pipe_paths   |~/flowr/pipelines        |
#> |flow_pipe_urls    |~/flowr/pipelines        |
#> |flow_platform     |local                    |
#> |flow_run_path     |~/flowr/runs             |
#> |local_cores       |4                        |
#> |module_cmds       |                         |
#> |time_format       |%a %b %e %H:%M:%S CDT %Y |
#> |verbose           |1                        |