Re-run a pipeline in case of hardware or software failures.

hardware no change required, simple rerun: rerun(x=flow_wd)

software either a change to flowmat or flowdef has been made: rerun(x=flow_wd, mat = new_flowmat, def = new_flowdef)

NOTE:

flow_wd: flow working directory, same input as used for status

rerun(x, ...)

# S3 method for character
rerun(x, ...)

# S3 method for flow
rerun(
  x,
  mat,
  def,
  start_from,
  samplename,
  execute = TRUE,
  kill = TRUE,
  select,
  ignore,
  verbose = opts_flow$get("verbose"),
  ...
)

Arguments

x
flow working directory

...
passed onto to_flow

mat
(optional) flowmat fetched from previous submission if missing. For more information regarding the format refer to to_flowmat

def
(optional) flowdef fetched from previous submission if missing. For more information regarding the format refer to to_flowdef

start_from
(required) which job to start from, this is a job name.

samplename
(optional) If flowmat contains multiple samples, provide the samplename, corresponding to the flow working directory provided.

execute
[logical] whether to execute or not

kill
(optional) logical indicating whether to kill the jobs from the previous execution of flow.

select
(optional) select a subset of jobs to rerun [character vector]

ignore
(optional) ignore a subset of jobs to rerun [character vector]

verbose
A numeric value indicating the amount of messages to produce. Values are integers varying from 0, 1, 2, 3, .... Please refer to the verbose page for more details. opts_flow$get("verbose")

Details

This function fetches details regarding the previous execution from the flow working directory (flow_wd).

It reads the flow object from the flow_details.rds file, and extracts flowdef and flowmat from it using to_flowmat and to_flowdef functions.

Using new flowmat OR flowdef for re-run:

Optionally, if either of flowmat or flowdef are supplied; supplied ones are used instead of those extracted from previous submission.

This functions efficiently updates job details of the latest submission into the previous file; thus information regarding previous job ids and their status is not lost.

Examples

if (FALSE) {

# 
rerun(wd = wd, fobj = fobj, execute = TRUE, kill = TRUE)

}