Monitor status of flow(s)

Summarize status of a flow OR multiple flows OR a high-level summary of all flows in a folder.

status(
  x,
  use_cache = FALSE,
  verbose = opts_flow$get("verbose"),
  out_format = "markdown",
  ...
)

get_status(x, ...)

# S3 method for flow
get_status(x, verbose, use_cache, out_format, ...)

# S3 method for character
get_status(x, verbose, use_cache, out_format, ...)

# S3 method for data.frame
get_status(x, verbose, use_cache, progress = TRUE, ...)

Arguments

x
path to the flow root folder or a parent folder to summarize several flows.

use_cache
This skips checking status of jobs which have already been completed a and assumes no new jobs were submitted in the flow(s) being monitored. [FALSE]

verbose
A numeric value indicating the amount of messages to produce. Values are integers varying from 0, 1, 2, 3, .... Please refer to the verbose page for more details. opts_flow$get("verbose")

out_format
passed onto knitr:::kable. supports: markdown, rst, html... [markdown]

...
not used

progress
Whether or not to show a progress bar, when fetching/reading files [TRUE]

Details

basename(x) is used in a wild card search.

Get status of all the flows: (all flows with 'sleep_pipe' in their name are checked and their status is shown) <br> flowr status x=~/flowr/runs/sleep_pipe*

Provide a high level summary of ALL flows in a folder: <br> flowr status x=~/flowr/runs

Use use_cache=TRUE to speed up checking the status. This assumes that no new jobs have been submitted and skips (re-)checking status of completed jobs.

Once all the jobs have been submitted to the cluster you may always use use_cache=TRUE.

Examples