Create a flowmat using a list a commands.

Create a flowmat (data.frame) using a named list a commands.

as.flowmat(): reads a file and checks for required columns. If x is data.frame checks for required columns.

to_flowmat(x, ...)

# S3 method for list
to_flowmat(x, samplename, ...)

# S3 method for data.frame
to_flowmat(x, ...)

# S3 method for flow
to_flowmat(x, ...)

as.flowmat(x, grp_col, jobname_col, cmd_col, ...)

is.flowmat(x)

Arguments

x
a named list, where name corresponds to the jobname and value is a vector of commands to run.

...
not used

samplename
character of length 1 or that of nrow(x) ['samplename']

grp_col
column used for grouping, default samplename.

jobname_col
column specifying jobname, default jobname

cmd_col
column specifying commands to run, default cmd

Examples

# Use this link for a few examples:
# http://flow-r.github.io/flowr/tutorial.html#define_modules

# create a flow mat, starting with a list of commands.
cmd_sleep = c("sleep 1", "sleep 2")
cmd_echo = c("echo 'hello'", "echo 'hello'")

# create a named list
lst = list("sleep" = cmd_sleep, "echo" = cmd_echo)
flowmat = to_flowmat(lst, samplename = "samp")

# read in a tsv; check and confirm format
ex = file.path(system.file(package = "flowr"), "pipelines")

flowmat = as.flowmat(file.path(ex, "sleep_pipe.tsv"))

#> > mat seems to be a file, reading it...

#> --> Using `samplename` as the grouping column

#> --> Using `jobname` as the jobname column

#> --> Using `cmd` as the cmd column

# if your column names are different than defaults, explicitly specify them.
flowmat = as.flowmat(file.path(ex, "sleep_pipe.tsv"), jobname_col = "jobname")

#> > mat seems to be a file, reading it...

#> --> Using `samplename` as the grouping column

#> --> Using `cmd` as the cmd column

# check if a object is a flowmat
is.flowmat(flowmat)

#> [1] TRUE

# create a flowdef, from this flowmat
flowdef = to_flowdef(flowmat)

#> Creating a skeleton flow definition

#> Following jobnames detected: sleep create_tmp merge size

#> --> checking if required columns are present...

#> --> checking if resources columns are present...

#> --> checking if dependency column has valid names...

#> --> checking if submission column has valid names...

#> --> checking for missing rows in def...

#> --> checking for extra rows in def...

#> --> checking submission and dependency types...

# create a flow object using flowmat and flowdef
fobj = to_flow(flowmat, flowdef)

#> > reading and checking flow def ...

#> --> checking if required columns are present...

#> --> checking if resources columns are present...

#> --> checking if dependency column has valid names...

#> --> checking if submission column has valid names...

#> --> checking for missing rows in def...

#> --> checking for extra rows in def...

#> --> checking submission and dependency types...

#> > Using flowname default: flowname

#> > Using flow_run_path default: ~/flowr/runs

#> > checking flow definition and flow matrix for consistency...

#> >        Detecting platform...

#> --> Will use platform from flow definition

#> >  working on... sample1

#> .

#> .

#> .

#> .

# extract a flowmat from a flow (here the samplename also contains the name of the flow)
flowmat2 = to_flowmat(fobj)

## submit the flow to the cluster (execute=TRUE) or do a dry-run (execute=FALSE)
if (FALSE) {
fobj2 = submit_flow(fobj, execute=FALSE)
fobj3 = submit_flow(fobj, execute=TRUE)

## Get the status or kill all the jobs
status(fobj3)
kill(fobj3)
}